aura/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic-settings==2.8.1
python-dotenv==1.0.1
psutil==7.0.0
pytest==8.3.5
